+++ packages/eyes-sdk-core/lib/sdk/EyesVisualGrid.js
@@ -2,6 +2,7 @@
 
 const BrowserType = require('../config/BrowserType')
 const Configuration = require('../config/Configuration')
+const GeneralUtils = require('../utils/GeneralUtils')
 const TypeUtils = require('../utils/TypeUtils')
 const ArgumentGuard = require('../utils/ArgumentGuard')
 const TestResultsFormatter = require('../TestResultsFormatter')
@@ -247,7 +248,7 @@ class EyesVisualGrid extends EyesCore {
       const snapshots = await this._takeDomSnapshots({breakpoints})
       const [{url}] = snapshots
       if (this.getCorsIframeHandle() === CorsIframeHandles.BLANK) {
-        snapshots.forEach(snapshot => CorsIframeHandler.blankCorsIframeSrcOfCdt(snapshot))
+        snapshots.forEach(CorsIframeHandler.blankCorsIframeSrcOfCdt)
       }
       // this._logger.verbose(`Dom extracted  (${checkSettings.toString()})   $$$$$$$$$$$$`)
 
@@ -291,36 +292,25 @@ class EyesVisualGrid extends EyesCore {
       const snapshot = await takeDomSnapshot({driver: this._driver})
       return Array(browsers.length).fill(snapshot)
     }
-    let widths = await Promise.all(
+    const widths = await Promise.all(
       browsers.map(async browser => {
-        if (TypeUtils.has(browser, 'width')) {
-          return browser.width
-        } else if (TypeUtils.has(browser, 'deviceName')) {
-          const {width} = await this._getEmulatedDeviceSize(browser)
-          return width
-        } else if (TypeUtils.has(browser, 'chromeEmulationInfo')) {
-          const {width} = await this._getEmulatedDeviceSize(browser.chromeEmulationInfo)
-          return width
-        } else if (TypeUtils.has(browser, 'iosDeviceInfo')) {
-          const {width} = await this._getIosDeviceSize(browser.iosDeviceInfo)
-          return width
-        }
+        const {width} = await this.getBrowserSize(browser)
+        return GeneralUtils.getBreakpointWidth(breakpoints, width)
       }),
     )
-    if (TypeUtils.isArray(breakpoints)) {
-      widths = widths.map(width => {
-        const breakpoint = breakpoints.find(breakpoint => breakpoint > width)
-        return breakpoint ? breakpoint - 1 : breakpoints[breakpoints.length - 1]
-      })
-    }
+    this._logger.log(
+      `taking multiple dom snapshots for widths: ${widths} (breakpoints=${breakpoints}`,
+    )
     const viewportSize = await this.getViewportSize()
     const snapshots = {}
     if (widths.includes(viewportSize.getWidth())) {
+      this._logger.log(`taking dom snapshot for existing width ${viewportSize.getWidth()}`)
       const snapshot = await takeDomSnapshot({driver: this._driver})
       snapshots[viewportSize.getWidth()] = snapshot
     }
     for (const width of widths) {
       if (snapshots[width]) continue
+      this._logger.log(`taking dom snapshot for width ${width}`)
       await this._driver.setViewportSize({width, height: viewportSize.getHeight()})
       const snapshot = await takeDomSnapshot({driver: this._driver})
       snapshots[width] = snapshot
@@ -385,22 +375,29 @@ class EyesVisualGrid extends EyesCore {
     return undefined
   }
 
-  async _getEmulatedDeviceSize({deviceName, screenOrientation = 'portrait'}) {
-    if (!this._emulatedDevicesSizesPromise) {
-      await this.getAndSaveRenderingInfo()
-      this._emulatedDevicesSizesPromise = this._serverConnector.getEmulatedDevicesSizes()
+  async getBrowserSize(browser) {
+    if (TypeUtils.has(browser, 'width')) {
+      return {width: browser.width, height: browser.height}
+    } else if (
+      TypeUtils.has(browser, 'chromeEmulationInfo') ||
+      TypeUtils.has(browser, 'deviceName')
+    ) {
+      const {deviceName, screenOrientation = 'portrait'} = browser.chromeEmulationInfo || browser
+      if (!this._emulatedDevicesSizesPromise) {
+        await this.getAndSaveRenderingInfo()
+        this._emulatedDevicesSizesPromise = this._serverConnector.getEmulatedDevicesSizes()
+      }
+      const devicesSizes = await this._emulatedDevicesSizesPromise
+      return devicesSizes[deviceName][screenOrientation]
+    } else if (TypeUtils.has(browser, 'iosDeviceInfo')) {
+      const {deviceName, screenOrientation = 'portrait'} = browser.iosDeviceInfo
+      if (!this._iosDevicesSizesPromise) {
+        await this.getAndSaveRenderingInfo()
+        this._iosDevicesSizesPromise = this._serverConnector.getIosDevicesSizes()
+      }
+      const devicesSizes = await this._iosDevicesSizesPromise
+      return devicesSizes[deviceName][screenOrientation]
     }
-    const devicesSizes = await this._emulatedDevicesSizesPromise
-    return devicesSizes[deviceName][screenOrientation]
-  }
-
-  async _getIosDeviceSize({deviceName, screenOrientation = 'portrait'}) {
-    if (!this._iosDevicesSizesPromise) {
-      await this.getAndSaveRenderingInfo()
-      this._iosDevicesSizesPromise = this._serverConnector.getIosDevicesSizes()
-    }
-    const devicesSizes = await this._iosDevicesSizesPromise
-    return devicesSizes[deviceName][screenOrientation]
   }
 }
 module.exports = EyesVisualGrid