+++ packages/eyes-sdk-core/test/unit/utils/GeneralUtils.spec.js
@@ -391,4 +391,15 @@ describe('GeneralUtils', () => {
       process.env.APPLITOOLS_MY_FLAG = undefined
     })
   })
+
+  describe('getBreakpointWidth()', () => {
+    it('works', () => {
+      assert.strictEqual(GeneralUtils.getBreakpointWidth([100], 50), 99)
+      assert.strictEqual(GeneralUtils.getBreakpointWidth([100], 150), 100)
+      assert.strictEqual(GeneralUtils.getBreakpointWidth([100, 300], 150), 299)
+      assert.strictEqual(GeneralUtils.getBreakpointWidth([100, 300], 500), 300)
+      assert.strictEqual(GeneralUtils.getBreakpointWidth([], 33), 33)
+      assert.strictEqual(GeneralUtils.getBreakpointWidth(null, 33), 33)
+    })
+  })
 })